Refactor code - implement stacked management
