+++ lib/main.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_shoe_app/views/cart/cart_view.dart';
-import 'package:flutter_shoe_app/views/homepage/homepage_view.dart';
-import 'package:flutter_shoe_app/views/shoeheader/shoeheader_view.dart';
+import 'package:flutter_shoe_app/views/home/home_view.dart';
+import 'package:flutter_shoe_app/views/home/shoe_header_view.dart';
 import 'package:flutter_shoe_app/views/shop/shop_view.dart';
 
-import 'views/homepage/homepage_view.dart';
+import 'views/home/home_view.dart';
 
 void main() {
   runApp(const MyApp());
@@ -45,7 +45,6 @@ class _MyHomePageState extends State<MyHomePage> {
         ),
         backgroundColor: Colors.black,
         shadowColor: Colors.blue,
-
       ),
 
       body: Align(
@@ -56,7 +55,7 @@ class _MyHomePageState extends State<MyHomePage> {
             ElevatedButton(
                 onPressed: () {
                   Navigator.push(context, MaterialPageRoute(builder: (context) {
-                    return const ShoeHeaderView();
+                    return ShoeCategory();
                   }));
                 },
                 child: const Text('Shoe Header')),
